+++ semester_5/lecture_27_hyperparameter_optimization/optuna/Algorithm.java
@@ -55,7 +55,7 @@ public class Algorithm {
         System.out.println("=".repeat(70));
         
         Algorithm algo = Algorithm.create();
-        float result = algo.suggest_float("", null, null);
+        int result = algo.suggest_float("", null, null);
         System.out.println("Result: " + result);
         System.out.println("=".repeat(70));
     }
